Moved files out of 'beta' folder
