+++ src/server/plugins/engine/index.ts
@@ -19,7 +19,7 @@ export {
   getFormContext,
   getFormModel,
   resolveFormModel
-} from '~/src/server/plugins/engine/beta/form-context.js'
+} from '~/src/server/plugins/engine/form-context.js'
 export * from '~/src/server/plugins/engine/form-availability.js'
 
 const globals = {
