+++ src/server/plugins/engine/routes/index.ts
@@ -10,11 +10,11 @@ import {
   EXTERNAL_STATE_APPENDAGE,
   EXTERNAL_STATE_PAYLOAD
 } from '~/src/server/constants.js'
-import { resolveFormModel } from '~/src/server/plugins/engine/beta/form-context.js'
 import {
   FormComponent,
   isFormState
 } from '~/src/server/plugins/engine/components/FormComponent.js'
+import { resolveFormModel } from '~/src/server/plugins/engine/form-context.js'
 import {
   checkFormStatus,
   findPage,
